Work around nytrocli /c${something}/eall show bogus error
nytrocli raises a bogus error if one attempts to enumerate the enclosures
of Nytro WarpDrive controller:

{
"Controllers":[
{
        "Command Status" : {
                "Controller" : 1,
                "Status" : "Failure",
                "ErrCd" : 1001,
                "Description" : "Un-supported command"
        }
}
]
}

As a work around don't query enclosures if the controller in question looks
like a Nytro WarpDrive (check the "Product Name" field).

The real solution is to fix nytrocli to give an empty list instead of a (bogus)
error.