Work around the nytrocli bug regarding the invalid JSON
The start format command prints serveral irrelevant messages before the actual
JSON output.  Surprisingly Avago refuses to fix this bug. Hence this work
around: skip the output until the first left curly bracket. This solution is
very unreliable and should not be used in production (for one, it breaks if
the debugging message contains the '{' character itself).
But at least we are able to run the functional tests and benchmarks.