Storcli: display the controller health information
3 parameters are available: 'overall_health', 'temperature',
'warranty_remaining'. The values are not interpreted in any way
and are passed to the caller as is.

Note: at the moment nytrocli raises a bogus error when the health
info of a MegaRAID controller is requested:

{
"Controllers":[
{
	"Command Status" : {
		"Controller" : 0,
		"Status" : "Failure",
		"ErrCd" : 1,
		"Description" : "None"
	},
	"Response Data" : {
		"Controller Health Info" : {
			"TemperatureROC" : 74,
			"Warranty Remaining" : 100,
			"Overall Health" : "GOOD",
			"Reason Code" : 0
		}
	}
}
]
}

Such output is silently ignored (no health information for such a
controller is provided).